Prototype diff formatting for discussion
Feature request:

It would be good to be able to see not only that actual and expected results are different,
but also one would like to see in which way they differ.

I couldnt find a way to extend leo for diffing in plugin-kind of way. So I started prototyping.

This commit prototypes a possible implementation with the following properties
* Because the relevant different depends on the type of matcher,
  each matcher supporting diffs (see SameMatcher) gets a $differ method to be called by the formatter
* also the matcher adds a "diff" to the format template vars
* on match, the abstract matcher adds the differ to match the result iff the result is unexpected
* Formatter reads and prints the $diff member from the matcher
* diffing is actually not implemented, that probably should be plugged in using a thirdparty lib

This indicates
* leo probably should not force diff calculation on users
* there should be a kind of extension point to add a differ implementation

So this bears the questions:
* is there a way already to hook in diffing?
* is there a better place than Matcher/Match/Formatter ?
** should diff calculation happen when matching or when formatting
* where/how would I register a Differ?
* should the Differ be added as mix-in on the Matchers?